refactor: 도메인 package 분류 작업
