+++ src/main/java/step3/domain/strategy/RandomMovingStrategy.java
@@ -1,4 +1,4 @@
-package step3.model;
+package step3.domain.strategy;
 
 import step3.RandomNumberProvider;
